we are now reading a server.ts file - with some hacks, but it works
