+++ nodemon.json
@@ -1,6 +1,6 @@
 {
   "ignore": [".git", "node_modules", "dist"],
   "watch": ["./src"],
-  "exec": "node src/server.js",
-  "ext": "js"
+  "exec": "npx tsx src/server.ts",
+  "ext": "ts"
 }
