+++ package.json
@@ -25,10 +25,14 @@
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-router-dom": "6.27.0",
-    "sirv": "3.0.0"
+    "sirv": "3.0.0",
+    "node": "20.18.1",
+    "tsx": "4.19.2"
   },
   "devDependencies": {
     "@eslint/js": "9.13.0",
+    "@types/body-parser": "1.19.5",
+    "@types/compression": "1.7.5",
     "@types/express": "5.0.0",
     "@types/react": "18.3.12",
     "@types/react-dom": "18.3.1",
